fix(cms): validate 2FA before 2fa/setup
